// frontend/src/types/strategy.ts
export type TEngagement = {
  model_username: string;
  model_option: string;
};

export type ISTRATEGY = {
  id: string;
  name: string;
  tweet_status: string;
  tweet_max_day: string;
  tweet_initial_day: string;
  tweet_days2max: string;
  tweet_growth: string;
  tweet_fluctuation: string;
  tweet_date: string;
  follow_status: string;
  follow_max_day: string;
  follow_initial_day: string;
  follow_days2max: string;
  follow_growth: string;
  follow_fluctuation: string;
  follow_limit: string;
  follow_date: string;
  likes_status: string;
  likes_max_day: string;
  likes_initial_day: string;
  likes_days2max: string;
  likes_growth: string;
  likes_fluctuation: string;
  likes_date: string;
  bookmark_status: string;
  bookmark_max_day: string;
  bookmark_initial_day: string;
  bookmark_days2max: string;
  bookmark_growth: string;
  bookmark_fluctuation: string;
  bookmark_date: string;
  quote_status: string;
  quote_max_day: string;
  quote_initial_day: string;
  quote_days2max: string;
  quote_growth: string;
  quote_fluctuation: string;
  quote_date: string;
  comment_status: string;
  comment_max_day: string;
  comment_initial_day: string;
  comment_days2max: string;
  comment_growth: string;
  comment_fluctuation: string;
  comment_date: string;
  engagements?: TEngagement[];
};

export type ITSTRATEGY = {
  id?: string;
  name?: string;
  tweet_status?: string;
  tweet_max_day?: string;
  tweet_initial_day?: string;
  tweet_days2max?: string;
  tweet_growth?: string;
  tweet_fluctuation?: string;
  tweet_date?: string;
  follow_status?: string;
  follow_max_day?: string;
  follow_initial_day?: string;
  follow_days2max?: string;
  follow_growth?: string;
  follow_fluctuation?: string;
  follow_limit?: string;
  follow_date?: string;
  likes_status?: string;
  likes_max_day?: string;
  likes_initial_day?: string;
  likes_days2max?: string;
  likes_growth?: string;
  likes_fluctuation?: string;
  likes_date?: string;
  bookmark_status?: string;
  bookmark_max_day?: string;
  bookmark_initial_day?: string;
  bookmark_days2max?: string;
  bookmark_growth?: string;
  bookmark_fluctuation?: string;
  bookmark_date?: string;
  quote_status?: string;
  quote_max_day?: string;
  quote_initial_day?: string;
  quote_days2max?: string;
  quote_growth?: string;
  quote_fluctuation?: string;
  quote_date?: string;
  comment_status?: string;
  comment_max_day?: string;
  comment_initial_day?: string;
  comment_days2max?: string;
  comment_growth?: string;
  comment_fluctuation?: string;
  comment_date?: string;
  engagements?: TEngagement[];
};